# app/routers/choke_workflow_router.py
import json
import logging
from typing import Any, Dict, Optional

from fastapi import APIRouter, HTTPException, Request
from pydantic import BaseModel, ConfigDict, Field, field_validator

from services.choke_sequential_agent_workflow import (
    calculate_final_choke_costing_from_saved_outputs,
    calculate_from_real_outputs,
    get_bom_agent_configuration_health,
    get_bom_output,
    get_component_output,
    get_component_outputs,
    get_most_output,
    get_most_outputs,
    get_writeback_debug,
    get_workflow_debug,
    get_workflow_state,
    get_customer_input_resolution,
    save_bom_output,
    save_component_output,
    save_most_output,
    retry_most_work_package,
    retry_bom_agent,
    run_storage_self_test,
    test_bom_agent_trigger,
    start_real_choke_workflow,
    trigger_most_operations,
    trigger_next_component_costing,
    update_commercial_fields,
)
from services.project_data_paths import (
    CustomerInputFileNotFound,
    get_workflow_run_paths,
)
from services.choke_writeback_mcp_diagnostic import (
    get_bom_agent_capability_diagnostic,
    get_mcp_schema_fingerprints,
    get_writeback_mcp_connectivity_diagnostic,
)
from services.workspace_agent_trigger_diagnostic import (
    run_minimal_trigger_diagnostic,
)
from services.choke_financial_workflow import (
    calculate_saved_financial_plan,
    get_financial_model_audit,
    get_financial_readiness,
    get_saved_financial_result,
    save_financial_reference_comparison,
    solve_saved_selling_price,
)


router = APIRouter(prefix="/api/choke-workflow", tags=["Choke Sequential Workflow"])
logger = logging.getLogger(__name__)


class StartWorkflowRequest(BaseModel):
    input_file: str
    dry_run: bool = False
    workflow_request_id: str | None = None


class SaveBomOutputRequest(BaseModel):
    project_code: str
    product_id: str
    trigger_run_id: str
    raw_json: Dict[str, Any]


class TriggerStageRequest(BaseModel):
    project_code: str
    product_id: str
    dry_run: bool = False
    force: bool = False
    include_unconfirmed: bool = False
    requested_by: str = "api"


class RetryBomRequest(BaseModel):
    project_code: str
    product_id: str


class RetryMostWorkPackageRequest(BaseModel):
    project_code: str
    product_id: str
    work_package_id: str
    dry_run: bool = False


class TestBomAgentTriggerRequest(BaseModel):
    project_code: str
    product_id: str
    drawing_file_url: str
    drawing_reference: Optional[str] = None


class SaveComponentOutputRequest(BaseModel):
    project_code: str
    product_id: str
    component_id: str
    trigger_run_id: str
    raw_json: Dict[str, Any]


class SaveMostOutputRequest(BaseModel):
    project_code: str
    product_id: str
    work_package_id: str
    most_scope_id: str
    trigger_run_id: str
    raw_json: Dict[str, Any]


class CalculateRealOutputsRequest(BaseModel):
    project_code: str
    product_id: str
    unit_data: Dict[str, Any] | None = None
    result_mode: str = "firm"

    @field_validator("result_mode")
    @classmethod
    def validate_result_mode(cls, value):
        normalized = str(value or "firm").strip().lower()
        if normalized not in {"firm", "preliminary"}:
            raise ValueError("result_mode must be 'firm' or 'preliminary'")
        return normalized


class UpdateCommercialFieldsRequest(BaseModel):
    project_code: str
    product_id: str
    product: str | None = None
    product_name: str | None = None
    product_family: str | None = None
    part_number: str | None = None
    customer: str | None = None
    final_customer: str | None = None
    customer_delivery_zone: str | None = None
    annual_quantity: float | None = None
    currency: str | None = None
    quotation_currency: str | None = None
    target_price_currency: str | None = None
    purchasing_currency: str | None = None
    delivery_country: str | None = None
    delivery_city: str | None = None
    target_price: float | None = None
    sop_date: str | None = None
    drawing_reference: str | None = None

    @field_validator("project_code", "product_id", "part_number", mode="before")
    @classmethod
    def normalize_identifiers(cls, value):
        if value is None:
            return None
        return str(value).strip()


class FinancialPlanRequest(BaseModel):
    model_config = ConfigDict(extra="allow")

    project_code: str
    product_id: str
    mode: str = "preliminary"
    financial_inputs: Dict[str, Any] = Field(default_factory=dict)

    @field_validator("mode")
    @classmethod
    def validate_mode(cls, value):
        normalized = str(value or "preliminary").strip().lower()
        if normalized not in {"firm", "preliminary"}:
            raise ValueError("mode must be 'firm' or 'preliminary'")
        return normalized


class FinancialReferenceComparisonRequest(BaseModel):
    project_code: str
    product_id: str
    historical_values: Dict[str, Any] = Field(default_factory=dict)
    explanations: Dict[str, str] = Field(default_factory=dict)
    acceptance: Dict[str, bool] = Field(default_factory=dict)
    validation_owner: str | None = None


def _financial_inputs(request: FinancialPlanRequest) -> Dict[str, Any]:
    payload = request.model_dump(exclude={"project_code", "product_id", "financial_inputs"})
    return {**request.financial_inputs, **payload}


def _handle(callback):
    try:
        return callback()
    except CustomerInputFileNotFound as exc:
        raise HTTPException(status_code=404, detail=exc.details) from exc
    except ValueError as exc:
        raise HTTPException(status_code=400, detail=str(exc)) from exc
    except FileNotFoundError as exc:
        raise HTTPException(status_code=404, detail=str(exc)) from exc
    except Exception as exc:
        raise HTTPException(status_code=500, detail=str(exc)) from exc


def _trigger_failure_http_status(result: Dict[str, Any]) -> int | None:
    state = result.get("state") or {}
    bom = state.get("bom") or result.get("bom") or {}
    failure_statuses = {
        "trigger_request_failed",
        "bom_trigger_failed",
        "failed_retryable",
        "failed_non_retryable",
        "failed",
    }
    if (
        result.get("success") is not False
        and result.get("status") not in failure_statuses
        and state.get("bom_status") not in failure_statuses
        and bom.get("status") not in failure_statuses
        and bom.get("lifecycle_status") not in failure_statuses
    ):
        return None
    safe_error = bom.get("safe_error") or {}
    code = str(safe_error.get("code") or "")
    upstream_status = safe_error.get("http_status") or bom.get("upstream_http_status")
    if code == "workspace_agent_http_401" or upstream_status == 401:
        return 401
    if code == "workspace_agent_http_403" or upstream_status == 403:
        return 403
    if code == "bom_agent_configuration_missing":
        return 503
    if code == "workspace_agent_timeout" or upstream_status == 504:
        return 504
    if code in {
        "workspace_agent_temporarily_unavailable",
        "workspace_agent_rate_limited",
        "workspace_agent_network_error",
    }:
        return 503
    return 502


def _raise_trigger_failure(result: Dict[str, Any]) -> None:
    status_code = _trigger_failure_http_status(result)
    if status_code is None:
        return
    state = result.get("state") or {}
    safe_error = ((state.get("bom") or {}).get("safe_error") or {})
    raise HTTPException(
        status_code=status_code,
        detail={
            "code": safe_error.get("code") or "bom_agent_trigger_failed",
            "message": (
                safe_error.get("message")
                or result.get("message")
                or "BOM Workspace Agent trigger failed."
            ),
            "retryable": bool(safe_error.get("retryable")),
            "state": state,
        },
    )


@router.post("/start")
def start_workflow(request: Request, payload: StartWorkflowRequest):
    logger.info(
        "BOM workflow request received: %s",
        json.dumps({
            "event": "request_received",
            "input_file": payload.input_file,
            "dry_run": payload.dry_run,
            "workflow_request_id": payload.workflow_request_id,
        }),
    )
    result = _handle(lambda: start_real_choke_workflow(
        input_file=payload.input_file,
        dry_run=payload.dry_run,
        request_base_url=str(request.base_url),
        workflow_request_id=payload.workflow_request_id,
    ))
    _raise_trigger_failure(result)
    return result


@router.get("/bom-agent-health")
def bom_agent_health():
    return _handle(get_bom_agent_configuration_health)


@router.get("/bom-agent-capabilities")
def bom_agent_capabilities():
    return _handle(get_bom_agent_capability_diagnostic)


@router.get("/workspace-trigger-diagnostic")
def workspace_trigger_diagnostic():
    return _handle(run_minimal_trigger_diagnostic)


@router.post("/storage-self-test")
def storage_self_test():
    return _handle(run_storage_self_test)


@router.get("/status/{project_code}/{product_id}")
def workflow_status(project_code: str, product_id: str):
    return _handle(lambda: get_workflow_state(project_code, product_id))


@router.get("/debug/{project_code}/{product_id}")
def workflow_debug(project_code: str, product_id: str):
    return _handle(lambda: get_workflow_debug(project_code, product_id))


@router.get("/writeback-debug/{project_code}/{product_id}")
def workflow_writeback_debug(project_code: str, product_id: str):
    return _handle(lambda: get_writeback_debug(project_code, product_id))


@router.get("/mcp-diagnostic")
def mcp_connectivity_diagnostic():
    return _handle(get_writeback_mcp_connectivity_diagnostic)


@router.get("/mcp-schema-fingerprints")
def mcp_schema_fingerprints():
    return _handle(get_mcp_schema_fingerprints)


@router.get("/bom-output/{project_code}/{product_id}")
def bom_output(project_code: str, product_id: str):
    return _handle(lambda: get_bom_output(project_code, product_id))


@router.get("/customer-input-resolution/{project_code}/{product_id}")
def customer_input_resolution(project_code: str, product_id: str):
    return _handle(lambda: get_customer_input_resolution(project_code, product_id))


@router.post("/update-commercial-fields")
def update_workflow_commercial_fields(request: UpdateCommercialFieldsRequest):
    dump = getattr(request, "model_dump", request.dict)
    fields = dump(exclude={"project_code", "product_id"}, exclude_unset=True)
    return _handle(lambda: update_commercial_fields(
        project_code=request.project_code,
        product_id=request.product_id,
        fields=fields,
    ))


@router.post("/save-bom-output")
def save_bom(request: SaveBomOutputRequest):
    return _handle(lambda: save_bom_output(
        project_code=request.project_code,
        product_id=request.product_id,
        trigger_run_id=request.trigger_run_id,
        raw_json=request.raw_json,
    ))


@router.post("/retry-bom")
def retry_bom(request: RetryBomRequest):
    result = _handle(lambda: retry_bom_agent(
        project_code=request.project_code,
        product_id=request.product_id,
    ))
    state = result.get("state") or {}
    bom = state.get("bom") or result.get("bom") or {}
    logger.info(
        "BOM retry API result: %s",
        json.dumps({
            "project_code": request.project_code,
            "product_id": request.product_id,
            "success": result.get("success"),
            "workflow_status": state.get("status") or result.get("status"),
            "bom_status": state.get("bom_status") or result.get("bom_status"),
            "bom_lifecycle_status": bom.get("lifecycle_status"),
            "trigger_run_id": bom.get("trigger_run_id"),
            "upstream_http_status": bom.get("upstream_http_status"),
            "retryable": bom.get("retryable"),
            "response_shape": sorted(result),
        }),
    )
    _raise_trigger_failure(result)
    return result


@router.post("/test-bom-agent-trigger")
def test_bom_trigger(request: TestBomAgentTriggerRequest):
    return _handle(lambda: test_bom_agent_trigger(
        project_code=request.project_code,
        product_id=request.product_id,
        drawing_file_url=request.drawing_file_url,
        drawing_reference=request.drawing_reference,
    ))


@router.post("/trigger-components")
def trigger_components(request: TriggerStageRequest):
    return _handle(lambda: trigger_next_component_costing(
        project_code=request.project_code,
        product_id=request.product_id,
        dry_run=request.dry_run,
        force=request.force,
        include_unconfirmed=request.include_unconfirmed,
        requested_by=request.requested_by,
    ))


@router.get("/component-output/{project_code}/{product_id}/{component_id}")
def component_output(project_code: str, product_id: str, component_id: str):
    return _handle(lambda: get_component_output(project_code, product_id, component_id))


@router.get("/component-outputs/{project_code}/{product_id}")
def component_outputs(project_code: str, product_id: str):
    return _handle(lambda: get_component_outputs(project_code, product_id))


@router.post("/save-component-output")
def save_component(request: SaveComponentOutputRequest):
    return _handle(lambda: save_component_output(
        project_code=request.project_code,
        product_id=request.product_id,
        component_id=request.component_id,
        trigger_run_id=request.trigger_run_id,
        raw_json=request.raw_json,
    ))


@router.post("/trigger-most")
def trigger_most(request: TriggerStageRequest):
    return _handle(lambda: trigger_most_operations(
        project_code=request.project_code,
        product_id=request.product_id,
        dry_run=request.dry_run,
        force=request.force,
        requested_by=request.requested_by,
    ))


@router.post("/retry-most-work-package")
def retry_most(request: RetryMostWorkPackageRequest):
    return _handle(lambda: retry_most_work_package(
        project_code=request.project_code,
        product_id=request.product_id,
        work_package_id=request.work_package_id,
        dry_run=request.dry_run,
    ))


@router.get("/most-output/{project_code}/{product_id}/{work_package_id}")
def most_output(project_code: str, product_id: str, work_package_id: str):
    return _handle(lambda: get_most_output(project_code, product_id, work_package_id))


@router.get("/most-outputs/{project_code}/{product_id}")
def most_outputs(project_code: str, product_id: str):
    return _handle(lambda: get_most_outputs(project_code, product_id))


@router.post("/save-most-output")
def save_most(request: SaveMostOutputRequest):
    return _handle(lambda: save_most_output(
        project_code=request.project_code,
        product_id=request.product_id,
        work_package_id=request.work_package_id,
        most_scope_id=request.most_scope_id,
        trigger_run_id=request.trigger_run_id,
        raw_json=request.raw_json,
    ))


@router.post("/calculate-from-real-outputs")
def calculate_real_outputs(request: CalculateRealOutputsRequest):
    return _handle(lambda: calculate_from_real_outputs(
        project_code=request.project_code,
        product_id=request.product_id,
    ))


@router.post("/calculate-final")
def calculate_final_outputs(request: CalculateRealOutputsRequest):
    return _handle(lambda: calculate_final_choke_costing_from_saved_outputs(
        project_code=request.project_code,
        product_id=request.product_id,
        unit_data_override=request.unit_data,
        result_mode=request.result_mode,
    ))


@router.get("/final-result/{project_code}/{product_id}")
def get_final_result(project_code: str, product_id: str):
    if any(part in {"", ".", ".."} or "/" in part or "\\" in part for part in [project_code, product_id]):
        raise HTTPException(status_code=400, detail="Invalid project_code or product_id")
    path = (
        get_workflow_run_paths(project_code, product_id)["run_dir"]
        / "final_choke_costing_result.json"
    )
    if not path.exists():
        raise HTTPException(
            status_code=404,
            detail=(
                "Final Choke costing result not found. "
                "Call POST /api/choke-workflow/calculate-final after all agent outputs are saved."
            ),
        )
    try:
        return json.loads(path.read_text(encoding="utf-8"))
    except json.JSONDecodeError as exc:
        raise HTTPException(status_code=500, detail="Saved final result is not valid JSON") from exc


@router.get("/financial-model-audit")
def financial_model_audit():
    return {"status": "ok", "audit": get_financial_model_audit()}


@router.get("/financial-readiness/{project_code}/{product_id}")
def workflow_financial_readiness(project_code: str, product_id: str):
    return _handle(lambda: get_financial_readiness(project_code, product_id))


@router.post("/calculate-financial-plan")
def calculate_workflow_financial_plan(request: FinancialPlanRequest):
    return _handle(lambda: calculate_saved_financial_plan(
        request.project_code,
        request.product_id,
        _financial_inputs(request),
    ))


@router.post("/solve-selling-price")
def solve_workflow_selling_price(request: FinancialPlanRequest):
    return _handle(lambda: solve_saved_selling_price(
        request.project_code,
        request.product_id,
        _financial_inputs(request),
    ))


@router.get("/financial-result/{project_code}/{product_id}")
def workflow_financial_result(project_code: str, product_id: str):
    return _handle(lambda: get_saved_financial_result(project_code, product_id))


@router.post("/compare-financial-reference")
def compare_workflow_financial_reference(
    request: FinancialReferenceComparisonRequest,
):
    return _handle(lambda: save_financial_reference_comparison(
        request.project_code,
        request.product_id,
        request.historical_values,
        request.explanations,
        request.acceptance,
        request.validation_owner,
    ))